Added tests for an issue with confirm delete cancelation and CKEditor
