+++ tests/src/FunctionalJavascript/ParagraphsFeaturesJavascriptTestBase.php
@@ -3,7 +3,9 @@
 namespace Drupal\Tests\paragraphs_features\FunctionalJavascript;
 
 use Drupal\Core\Entity\Entity\EntityFormDisplay;
+use Drupal\editor\Entity\Editor;
 use Drupal\field\Entity\FieldStorageConfig;
+use Drupal\filter\Entity\FilterFormat;
 use Drupal\FunctionalJavascriptTests\WebDriverTestBase;
 use Drupal\paragraphs\Tests\Classic\ParagraphsCoreVersionUiTestTrait;
 use Drupal\Tests\paragraphs\FunctionalJavascript\LoginAdminTrait;
@@ -31,6 +33,7 @@ abstract class ParagraphsFeaturesJavascriptTestBase extends WebDriverTestBase {
     'field_ui',
     'link',
     'node',
+    'ckeditor',
     'paragraphs',
     'paragraphs_test',
     'paragraphs_features',
@@ -103,6 +106,42 @@ abstract class ParagraphsFeaturesJavascriptTestBase extends WebDriverTestBase {
     EntityFormDisplay::load('paragraph.test_nested.default')
       ->setComponent('field_paragraphs', $component)
       ->save();
+  }
+
+  /**
+   * Create CKEditor for testing of CKEditor integration.
+   */
+  protected function createEditor() {
+    // Create a text format and associate CKEditor.
+    $filtered_html_format = FilterFormat::create([
+      'format' => 'filtered_html',
+      'name' => 'Filtered HTML',
+      'weight' => 0,
+    ]);
+    $filtered_html_format->save();
+
+    Editor::create([
+      'format' => 'filtered_html',
+      'editor' => 'ckeditor',
+    ])->save();
+
+    // After createTestConfiguration, $this->admin_user will be created by
+    // LoginAdminTrait used in base class.
+    $this->admin_user->addRole($this->createRole(['use text format filtered_html']));
+    $this->admin_user->save();
+  }
+
+  /**
+   * Get CKEditor ID, that can be used to get CKEditor objects in JavaScript.
+   *
+   * @param int $paragraph_index
+   *   Text paragraph index.
+   *
+   * @return string
+   *   Returns Id for CKEditor.
+   */
+  protected function getCkEditorId($paragraph_index) {
+    return $this->getSession()->getPage()->find('xpath', '//*[@data-drupal-selector="edit-field-paragraphs-' . $paragraph_index . '"]//textarea')->getAttribute('id');
   }
 
 }
